Start fleshing out KDE Project dependencies.
This also demonstrates that comments work. :P

For those adding their own rules, please see the README for the current
spec. Note that wildcard-style rules are not currently supported by
kdesrc-build but if that is a large issue let me know and I can see
about adding support and making that "official" in the spec.

If you do add your own dependency rules please check with a simple
kdesrc-build --pretend run on your module(s) to make sure no exceptions
are thrown, as right now kdesrc-build will completely skip dependency
resolution if any part of the process throws an exception.