dependency-data: Add update for proposed kwebkitpart change.
There's a proposal on release-team mailing list [1] to move kwebkitpart
under kde/kde-baseapps. It sounds good, but if it goes through we need
to make sure kdesrc-build doesn't try to build kwebkitpart before
kde-baseapps (similar as to what we do with konsole and kate now).

If the proposal doesn't go through this is merely extraneous info so
it's safe to commit at this point either way.

[1] http://mail.kde.org/pipermail/release-team/2012-March/005531.html