tools: Support 'foo/bar*' wildcards in addition to 'foo/bar/*'.
The "spec" (such as it was) for logical-module-structure only supported
wildcarding entire path components, not sub-parts of a path component
(e.g.  'foo/bar*' wouldn't work).

The usage of logical-module-structure appears to have evolved past that;
CI at least supports a syntax like that, so let's adapt ourselves.

I've tested many modules but let me know if this broke something.