Remove khelpcenter as build dependency of kdevelop
Summary:
Seems KDevelop does not use anything from khelpcenter during the build
or the auto tests.

And at runtime only invokes (via KXMLGUI's default KHelpMenu)
the help:/ url, for which there could be (in theory) another
handler registered. But then runtime deps are in general not noted
in the build metadata, so there seems no reason to make an exception
here and also only for KDevelop, when khelpcenter is used for
other apps as well.

Reverts part of ed155055521ccca3b5849288a9ec336e59912579

Reviewers: #kdevelop, apol, bcooksley, kfunk

Reviewed By: #kdevelop, kfunk

Differential Revision: https://phabricator.kde.org/D6193