Break dependency cycle between ECM and Qt5.
Since kdesrc-build is gaining support for building Qt5, and a more
accurate dependency resolver for kdesrc-build is also in development, it
would be good for kdesrc-build to not have to special-case dependencies
involving Qt5.

kdesrc-build has magic to ignore Qt5 but can't remove that magic until
the dependency cycle between Qt5 and extra-cmake-modules is broken. This
is done by negating the dependency of Qt5 onto extra-cmake-modules, as
reviewed by @bcooksley.

Fixes kde/kdesrc-build#22