Remove duplicate project entries.
Some JSON parsers don't accept JSON files with duplicate keys in the
same dict. Notably Mojo::JSON (or whatever that is using behind the
scenes).

Luckily, all of the duplicates contained the same values themselves so this
shouldn't alter anything.

To easily find duplicates again in the future, this is the command I'd
run:

    egrep -o '"[^"]+/[^"]+": +{' logical-module-structure | \
    sed -e 's/:.*$//' | sort | uniq -d

This looks for entries that look like they are project dicts, strips
them down to just the quoted key, and checks for duplicate keys.