plasma-thunderbolt has no Plasma/5.16 branch
(this file said otherwise, so kdesrc-build gave an error, when using
stable-kf5-qt5)

I suppose that's on purpose because it wasn't released yet?
Will it be part of the next Plasma release and then get a stable branch?
(in that case, please remember to revert this commit, so that it uses
the default settings for kde/workspace/* again)

CCMAIL: jr@jriddell.org, dvratil@kde.org